+++ conf.d/04-server_contact_info.cfg.lua
@@ -1,8 +1,10 @@
+local stringy = require "stringy" 
+
 contact_info = {
-  abuse = { os.getenv("SERVER_CONTACT_INFO_ABUSE") };
-  admin = { os.getenv("SERVER_CONTACT_INFO_ADMIN") };
-  feedback = { os.getenv("SERVER_CONTACT_INFO_FEEDBACK") };
-  sales = { os.getenv("SERVER_CONTACT_INFO_SALES") };
-  security = { os.getenv("SERVER_CONTACT_INFO_SECURITY") };
-  support = { os.getenv("SERVER_CONTACT_INFO_SUPPORT") };
+  abuse = stringy.split(os.getenv("SERVER_CONTACT_INFO_ABUSE"), ", ");
+  admin = stringy.split(os.getenv("SERVER_CONTACT_INFO_ADMIN"), ", ");
+  feedback = stringy.split(os.getenv("SERVER_CONTACT_INFO_FEEDBACK"), ", ");
+  sales = stringy.split(os.getenv("SERVER_CONTACT_INFO_SALES"), ", ");
+  security = stringy.split(os.getenv("SERVER_CONTACT_INFO_SECURITY"), ", ");
+  support = stringy.split(os.getenv("SERVER_CONTACT_INFO_SUPPORT"), ", ");
 }